start run reoptNature_no_DIC_GM15_npp1
reoptimize the nature paper but set the dicscale to 0.

questions
1. do we get a better fit to O2 obs when we exclude DIC from objective function?
2. do we get a reasonable C:P uptake pattern when we exclude DIC from objective function?
  - if very variable, then O2:P might need to be more variable than C:P in the model. in which case, the oxygen model is not what is limiting the variability of C:P.
  - if constant, then maybe the O2:P ratio needs to be more constant than the O2:C ratio

goal: testing if theres a trade-off between the model fit to O2 and the model fit to DIC. since the O2:C ratio is globally uniform in the model, which doesn't allow for any decoupling of the biological carbon and oxygen cycles. We know the respiration quotient can vary to a some extent, so maybe the lack of a variable O:C ratio makes it so the model cannot always match both the DIC and O2 at the same time.

Also because I'm not sure if model parameters are sensitive to anthropogenic carbon correction. Currently, the model loads the same DIC field corrected for anthropogenic carbon, regardless of alterations to the biogeochemical model. The corrected DIC is calculated from running the BGC model, using a prior iteration of the Nature paper optimal parameters, forward in time, and removing the calculated additional DICant timeseries from the observations of DIC.

The problem with using DIC observations in the objective function is that ocean DIC has changed significantly over the course of the observational record. Since we are trying to solve for the preindustrial steady state, we need to remove the anthropogenic carbon signal. In order to remove the signal, we need a model of ocean carbon uptake. The biological component of this anthropogenic carbon uptake will change depending on the biogeochemical model parameters.
I assume this change is small, since the solubility pump contributes much more to carbon intrusion into the deep ocean than the biological carbon pump does.
However, it is possible that this assumption is not true, and that the model optimization is sensitive to the parameters used when calculating the transient anthropogenic carbon uptake signal.
by excluding DIC from the objective function, we can avoid this potential problem (with the downside being that this removes a very large amount of data used to constrain the model)