SetPar: change to gamma= 0.4; start run reoptNature_with_dop_GM15_npp1_gamma40pct.out
Goal:
To see how C:P varies as a funciton ofsequestration time if we allow for both labile P and C. the previous version only had labile C, but no labile P, so there was a lot of fast remineralization of carbon, making C:P decrease as a funciton of sequestration time. but this goes against the idea of preferential remineralization of phosporus. I want to see if our result still holds when we have a significant amount of labile DOP.

Description:
- re-optimize the BGC model from the Nature paper (Wang et al., 2023), but include DOP in the objective function, and scale both DOC and DOP equally. Turn on gamma in SetPar. so gamma = 0.4; fraction of DOP production that goes to the labile pool.
- SetUp is same as reoptNature_GM15_npp1_CbPM. using CbPM NPP and GM15 p2c for satellite NPP conversion. load inital guess on CO and params from hojongs previous run. Temp and DIP_obs are smoothed.
- load initial parameters from = ../output/reoptNature_with_dop_GM15_npp1_CTL_He_PCO_DIP1e+00_DIC1e+00_DOC1e+00_ALK1e+00_O21e+00_xhat.mat