neglogpost: add increment iter inside resetpar check
Previously, iter only increased when after solver found good values and continued running neglogpost beyond the ResetPar check.
This update makes it so that the fb variable inside ResetPar increases with each call to ResetPar.
In a previous run, the solver suggested suboptimal values for 10 consecutive iterations and then terminated without finding good values  or running the model.
The goal of this change is to relax constraints on the realistic range of parameter values, should help useful when initial values are very far from the optimal set.