SetPar: change to gamma = 0.01; start run reoptNature_with_dop_GM15_npp1_gamma1pct
same as reoptNature_with_dop_GM15_npp1_gamma40pct but now, reduce gamma to 1%

choosing a small value because reoptNature_with_dop_GM15_npp1 found sigP = 0.027. so gamma = 0.01 would be roughly half of the total fraction of production routed to the dissolved pool.
(in eqPcycle, looks like gamma is a fraction of the total production, not a fraction of just the DOP portion of production)