SetPar: set gamma = 0.3; start run reoptNature_with_dop_GM15_npp1_gamma30pct
- set labile DOP = 30% of P production.
- This time, don't initialize from my previous optimization. use the initCO file Wei-Lei used in the Nature paper, and default parameters in SetPar.

To Do:
Will start this run once there is enough memory on greenplanet.
in the meantime, check for any errors in eqPcycle and eqCcyle when gamma is not zero: make sure the terms with gamma are in the gradient and hessian equations

In previous run with gamma = 0.4, Ccycle never converged. The run with gamma = 0.2 has also been stuck in eqPcycle for a few hours, and returned warning
[Warning: Matrix is close to singular or badly scaled. Results may be inaccurate. RCOND =  6.828150e-29.]
so large gamma may be a problem.

ideas:
Maybe DOPl should be fueling POP production too. because we do not allow dop to be taken up as a nutrient source for growth, maybe there isnt enough phosphate to support a lot of short lived DOP production in addition to all the POP and DOP production needed to get enough export to match the observations