(LOOK) neglogpost: add DOPl to DOP in objective function
After solving P cycle, add together DOP and DOPl before calculating error.
>> DOP = DOP + DOPl; % sum of semilabile and labile DOP ;

When computing objective function gradient,
change:
>> opx  = Px(2*nwet+1:end ,:)
to:
>> opx  = Px(2*nwet+1:3*nwet,:) + Px(3*nwet+1:4*nwet,:);

Note: previously, opx = Px(2*nwet+1:end,:) was twice as long as nwet. because of range @*nwet+1:end, this was the vector corresponding to DOP followed by DOPl. This didn't cause an error because in fx equation, only the indices idop get extracted from opx. idop is always values <nwet, so all the values corresponding to DOPl were ignored.

When calculating objective funciton Hessian,
change:
>> opxx = Pxx(2*nwet+1 : 3*nwet, :)
to
>> opxx = Pxx(2*nwet+1 : 3*nwet, :) + Pxx(3*nwet+1:4*nwet,:);

This change makes DOP more consistent with how DOC is treated in the objective function. However, this change assumes our refractory correction from the DOPobs is good and that the remaining DOP is the sum of DOP and DOPl.
So it might actually be better not to have DOPl in the objective function, and say the labile component of DOP is implicitly accounted for in our DOPobs correction (when we removed an estimate of the refractory component of DOP from the DOP observations)